pyproject.toml
[tool.poetry]
name = "deep_blueprint"
version = "0.1.0"
description = ""
authors = ["Niels Sayez <niels.sayez@uclouvain.be>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<=3.10.4"
# python = ">=3.7.1,<=3.8.10"
albumentations = "1.2.1"
cachetools = "5.2.0"
click = "8.1.3"
filelock = "3.7.1"
ipython = "^7.5.0"
ipywidgets = "^7.7.1"
kornia = "0.6.6"
matplotlib = "3.5.2"
monai = "^0.8.0"
# numba = "0.56.0"
omegaconf = "^2.1.1"
pandas = "1.3.5"
opencv-python = "^4.6.0"
pytorch_lightning = "1.6.5"
scikit-learn = "1.1.1"
scipy = "^1.7.0"
seaborn = "^0.11.1"
torch = "1.11.0"
torchmetrics = "^0.5.0"
torchvision = "^0.12.0"
tqdm = "4.64.0"
# typer = "0.6.1"
scikit-image = "^0.19.2"
hydra-core =  "^1.1.1"
hydra-submitit-launcher = "^1.1.6"
numpy = "^1.23.1"
wandb = "^0.12.21"
mlflow = "^1.27.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
hypothesis = {extras = ["numpy"], version = "^5.41.2"}
jupyterlab = "<3.0.0"
ipympl = "^0.5.7"
black = "^19.10b0"
jupyterlab-git = "^0.22.2"
mypy = "^0.790"
kaggle = "^1.5.10"
jupytext = "^1.11.3"

[tool.poetry.scripts]
deep_blueprint = 'deep_blueprint.train:main'

[tool.pytest.ini_options]
log_cli = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
